+++ person.rb
@@ -7,7 +7,7 @@ class Person < Nameable
 
   def initialize(age, name = 'Unknown', parent_permission: true)
     super()
-    @id = Random.rand(1..1000)
+    @id = Random.rand(1..10_000)
     @name = name
     @age = age
     @parent_permission = parent_permission
