+++ rental.rb
@@ -1,16 +1,16 @@
-require_relative 'person'
 require_relative 'book'
+require_relative 'person'
 
 class Rental
-  attr_accessor :date, :person, :book
+  attr_accessor :date, :book, :person
 
-  def initialize(date, person, book)
+  def initialize(date, book, person)
     @date = date
-
-    @person = person
-    person.rentals << self
 
     @book = book
     book.rentals << self
+    
+    @person = person
+    person.rentals << self
   end
 end
